requirements.txt
openai>=0.27.0
requests>=2.28.0
PyYAML>=6.0
GitPython>=3.1.31
flask>=2.2.5
python-dotenv>=1.0.0
typing-extensions>=4.0.1
